Move and rename class and method
